Split DetailviewExtensionHook
refs #2152
