+++ modules/monitoring/library/Monitoring/Hook/DetailviewExtensionHook.php
@@ -14,22 +14,8 @@ use Icinga\Web\View;
  *
  * Extend this class if you want to extend the detail view of monitored objects with custom HTML.
  */
-abstract class DetailviewExtensionHook
+abstract class DetailviewExtensionHook extends BaseViewExtensionHook
 {
-    /**
-     * The view the generated HTML will be included in
-     *
-     * @var View
-     */
-    private $view;
-
-    /**
-     * The module of the derived class
-     *
-     * @var Module
-     */
-    private $module;
-
     /**
      * Create a new hook
      *
@@ -55,58 +41,4 @@ abstract class DetailviewExtensionHook
      * @return  string
      */
     abstract public function getHtmlForObject(MonitoredObject $object);
-
-    /**
-     * Get {@link view}
-     *
-     * @return View
-     */
-    public function getView()
-    {
-        return $this->view;
-    }
-
-    /**
-     * Set {@link view}
-     *
-     * @param   View $view
-     *
-     * @return  $this
-     */
-    public function setView($view)
-    {
-        $this->view = $view;
-        return $this;
-    }
-
-    /**
-     * Get the module of the derived class
-     *
-     * @return Module
-     */
-    public function getModule()
-    {
-        if ($this->module === null) {
-            $class = get_class($this);
-            if (ClassLoader::classBelongsToModule($class)) {
-                $this->module = Icinga::app()->getModuleManager()->getModule(ClassLoader::extractModuleName($class));
-            }
-        }
-
-        return $this->module;
-    }
-
-    /**
-     * Set the module of the derived class
-     *
-     * @param Module $module
-     *
-     * @return $this
-     */
-    public function setModule(Module $module)
-    {
-        $this->module = $module;
-
-        return $this;
-    }
 }
